Refactor component and widgets structures from home module to core
