+++ lib/app/core/widgets/explanation_widget.dart
