+++ lib/app/core/widgets/zoomable_image_widget.dart
